backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Config
pydantic-settings==2.7.1
pydantic[email]

# OCR - Google Document AI
google-cloud-documentai==2.34.0

# LLM - Groq
groq==0.15.0

# Supabase
supabase==2.11.0

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
